add edit feature to cabin

+++ src/services/apiCabins.ts
@@ -12,47 +12,55 @@ export async function getCabins(): Promise<Cabin[]> {
   return cabins;
 }
 
-export async function createCabin(newCabin: NewCabin): Promise<Cabin> {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
-    '/',
-    '',
-  );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+export async function createEditCabin(newCabin: NewCabin, id?: number): Promise<Cabin> {
+  const hasImagePath = typeof newCabin.image === 'string';
 
-  // 1. Create Cabin
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
-    .select()
-    .single();
+  const imageName =
+    typeof newCabin.image === 'object' && 'name' in newCabin.image
+      ? `${Math.random()}-${newCabin.image.name}`.replaceAll('/', '')
+      : '';
+
+  const imagePath = hasImagePath
+    ? (newCabin.image as string)
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+  const query = id
+    ? supabase
+        .from('cabins')
+        .update({ ...newCabin, image: imagePath })
+        .eq('id', id)
+        .select()
+        .single()
+    : supabase
+        .from('cabins')
+        .insert([{ ...newCabin, image: imagePath }])
+        .select()
+        .single();
+
+  const { data, error } = await query;
 
   if (error) {
     console.error(error);
     throw new Error('Cabin could not be created');
   }
 
-  // 2. Upload image
-  const { error: storageError } = await supabase.storage
-    .from('cabin-images')
-    .upload(imageName, newCabin.image);
+  if (!hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from('cabin-images')
+      .upload(imageName, newCabin.image);
 
-  // 3. Delete cabin if there was an error uploading image
-  if (storageError) {
-    await supabase.from('cabins').delete().eq('id', data.id);
-    console.error(storageError);
-    throw new Error(
-      'Cabin image could not be uploaded and the cabin was not created',
-    );
+    if (storageError) {
+      await supabase.from('cabins').delete().eq('id', data.id);
+      console.error(storageError);
+      throw new Error('Cabin image could not be uploaded and the cabin was not created');
+    }
   }
 
   return data;
 }
 
 export async function deleteCabin(cabinId: number): Promise<Cabin> {
-  const { data: cabin, error } = await supabase
-    .from('cabins')
-    .delete()
-    .eq('id', cabinId);
+  const { data: cabin, error } = await supabase.from('cabins').delete().eq('id', cabinId);
 
   if (error) {
     console.error(error);
